# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Issue Tracker API
  version: 1.0.0
servers:
  - url: http://localhost:8000
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: User created
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /auth/login:
    post:
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '201':
          description: User authenticated
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user
      security:
        - accessTokenCookie: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout current user
      security:
        - accessTokenCookie: []
      responses:
        '201':
          description: Cookie cleared
        '401':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: access-token
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          oneOf:
            - type: string
            - type: array
              items: { type: string }
      required: [message]
    AuthUser:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
      required: [id, email]
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        password: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, email]
...
    RegisterDto:
      type: object
      properties:
        email: { type: string, format: email }
        password:
          type: string
          minLength: 6
          maxLength: 22
      required: [email, password]
    LoginDto:
      type: object
      properties:
        email: { type: string, format: email }
        password:
          type: string
          minLength: 6
          maxLength: 22
      required: [email, password]
...